fix(sdk-go): make the Go SDK publishable — correct tag scheme + green build (#674)
The Go SDK has never published despite having a workflow, because the publish
mechanism was wired wrong and the module didn't fully pass `go test`/`go vet`.

Workflow (go-publish-sdk.yml):
- Tag trigger `go-sdk-v*` -> `sdk/go/v*`. This module lives in a subdirectory
  (github.com/SaaSy-Solutions/mockforge/sdk/go), so Go's versioning rules
  require the tag to be `sdk/go/v0.1.0`; `go get .../sdk/go@v0.1.0` only ever
  resolves a `sdk/go/v0.1.0` tag. `go-sdk-v*` could never make it installable.
- Fix the proxy warm-up URL: `saa%21sy-solutions` (percent-encoded, and missing
  two capitals — resolves to the wrong-cased `saaSy-solutions`) -> the correct
  module-path escaping `!saa!sy-!solutions`, which resolves to the canonical
  `SaaSy-Solutions` matching go.mod. Validated against the live proxy.
- `go vet ./...` -> `go vet -unsafeptr=false ./...`: the WASM plugin sub-package
  reads guest linear memory via the canonical unsafe.Slice/uintptr ABI pattern;
  unsafeptr is a false positive there. Every other analyzer stays on.
- Test step no longer swallows failures (`|| echo`) now that the suite is green.

Module:
- Split the integration tests into mockserver_integration_test.go with a proper
  top-of-file `//go:build integration` constraint. They were appended to
  mockserver_test.go with the constraint placed mid-file (invalid — must precede
  the package clause), which broke `go test ./...` for everyone. Runtime
  t.Skip() behaviour is unchanged.
- gofmt the WASM plugin file (1 line) so the published module is fmt-clean.

After merge, tag `sdk/go/v0.1.0` to publish (Go needs no registry secret).
